+++ src-docs/src/views/combo_box/combo_box.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 
 import { EuiComboBox } from '../../../../src/components';
-import { DisplayToggles } from '../form_controls/shared';
+import { DisplayToggles } from '../form_controls/display_toggles';
 
 export default class extends Component {
   constructor(props) {
