+++ src-docs/src/views/super_select/super_select.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 
 import { EuiSuperSelect } from '../../../../src/components';
-import { DisplayToggles } from '../form_controls/shared';
+import { DisplayToggles } from '../form_controls/display_toggles';
 
 export default class extends Component {
   constructor(props) {
